对blog-admin的接口进行了调整，新增typeMapper、typeMapping.xml,同理tag也新增了mapper，controller也对tag、type进行了新增
