+++ blog-admin-6001/src/main/java/com/eva/mapper/AdminMapper.java
@@ -1,7 +1,6 @@
 package com.eva.mapper;
 
 import com.eva.dto.Tag;
-import com.eva.dto.Type;
 import com.eva.dto.User;
 import org.springframework.stereotype.Repository;
 
@@ -16,24 +15,4 @@ public interface AdminMapper {
     User findByUsernameAndPassword(Map map);
 
     User getUser(User user);
-
-    int addType(Type type);
-
-    List<Type> getTypesByPage();
-
-    int updateType(Type type);
-
-    Type getTypeByTypeId(String typeId);
-
-    int deleteTypeByTypeId(Type type);
-
-    int addTag(Tag tag);
-
-    List<Tag> getTagsByPage();
-
-    int updateTag(Tag tag);
-
-    Tag getTagByTagId(Tag tag);
-
-    int deleteTagByTagId(Tag tag);
 }
