+++ blog-admin-6001/src/main/resources/mapping/AdminMapping.xml
@@ -2,23 +2,11 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.eva.mapper.AdminMapper">
 
-    <resultMap id="BaseTypeMap" type="com.eva.dto.Type">
-        <id column="type_id" jdbcType="INTEGER" property="typeId" />
-        <result column="type_name" jdbcType="VARCHAR" property="typeName" />
-        <result column="delete_flag" jdbcType="VARCHAR" property="deleteFlag"/>
-    </resultMap>
 
-    <resultMap id="BaseTagMap" type="com.eva.dto.Tag">
-        <id column="tag_id" jdbcType="VARCHAR" property="tagId"/>
-        <result column="tag_name" jdbcType="VARCHAR" property="tagName"/>
-        <result column="delete_flag" jdbcType="VARCHAR" property="deleteFlag"/>
-    </resultMap>
-    <update id="updateType" parameterType="com.eva.dto.Type">
-        update type set type_name = #{typeName} where type_id = #{typeId}
-    </update>
-    <delete id="deleteTagByTagId" parameterType="com.eva.dto.Tag">
-        update tag set delete_flag = '1' where tag_id = #{tagId}
-    </delete>
+
+
+
+
 
     <select id="selectUser" resultType="com.eva.dto.User">
         select * from user
@@ -32,33 +20,6 @@
         select username from user where username=#{username};
     </select>
 
-    <select id="getTypesByPage" resultMap="BaseTypeMap">
-        select type_id,type_name from type where delete_flag = 0
-    </select>
-    <select id="getTypeByTypeId" parameterType="com.eva.dto.Type" resultMap="BaseTypeMap">
-        select * from type where type_id = #{typeId} and delete_flag = '0'
-    </select>
-    <select id="getTagsByPage" resultMap="BaseTagMap">
-        select * from tag where delete_flag = 0
-    </select>
 
-    <select id="getTagByTagId" parameterType="com.eva.dto.Tag" resultMap="BaseTagMap">
-        select * from tag where tag_id = #{tagId} and delete_flag = '0'
-    </select>
 
-    <insert id="addType" parameterType="com.eva.dto.Type">
-        insert into type values (#{typeId},#{typeName},'0');
-    </insert>
-
-    <insert id="addTag" parameterType="com.eva.dto.Tag">
-        insert into tag values (#{tagId},#{tagName},'0');
-    </insert>
-
-    <update id="deleteTypeByTypeId" parameterType="com.eva.dto.Type">
-        update type set delete_flag = '1' where type_id = #{typeId}
-    </update>
-
-    <update id="updateTag" parameterType="com.eva.dto.Tag">
-        update tag set tag_name = #{tagName} where tag_id = #{tagId}
-    </update>
 </mapper>
