+++ blog-admin-6001/src/main/resources/mapping/TagMapping.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.eva.mapper.TagMapper">
+
+    <resultMap id="BaseTagMap" type="com.eva.dto.Tag">
+        <id column="tag_id" jdbcType="VARCHAR" property="tagId"/>
+        <result column="tag_name" jdbcType="VARCHAR" property="tagName"/>
+        <result column="delete_flag" jdbcType="VARCHAR" property="deleteFlag"/>
+    </resultMap>
+
+    <select id="getTagsByPage" resultMap="BaseTagMap">
+        select * from tag where delete_flag = 0
+    </select>
+
+    <select id="getTagByTagId" parameterType="com.eva.dto.Tag" resultMap="BaseTagMap">
+        select * from tag where tag_id = #{tagId} and delete_flag = '0'
+    </select>
+
+
+
+    <insert id="addTag" parameterType="com.eva.dto.Tag">
+        insert into tag values (#{tagId},#{tagName},'0');
+    </insert>
+
+
+
+    <update id="updateTag" parameterType="com.eva.dto.Tag">
+        update tag set tag_name = #{tagName} where tag_id = #{tagId}
+    </update>
+
+    <delete id="deleteTagByTagId" parameterType="com.eva.dto.Tag">
+        update tag set delete_flag = '1' where tag_id = #{tagId}
+    </delete>
+</mapper>
