+++ blog-admin-6001/src/main/resources/mapping/TypeMapping.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.eva.mapper.TypeMapper">
+
+    <resultMap id="BaseTypeMap" type="com.eva.dto.Type">
+        <id column="type_id" jdbcType="INTEGER" property="typeId" />
+        <result column="type_name" jdbcType="VARCHAR" property="typeName" />
+        <result column="delete_flag" jdbcType="VARCHAR" property="deleteFlag"/>
+    </resultMap>
+
+    <insert id="addType" parameterType="com.eva.dto.Type">
+        insert into type values (#{typeId},#{typeName},'0');
+    </insert>
+
+    <update id="updateType" parameterType="com.eva.dto.Type">
+        update type set type_name = #{typeName} where type_id = #{typeId}
+    </update>
+
+    <select id="getTypesByPage" resultMap="BaseTypeMap">
+        select type_id,type_name from type where delete_flag = 0
+    </select>
+
+    <select id="getTypeByTypeId" parameterType="com.eva.dto.Type" resultMap="BaseTypeMap">
+        select * from type where type_id = #{typeId} and delete_flag = '0'
+    </select>
+
+    <update id="deleteTypeByTypeId" parameterType="com.eva.dto.Type">
+        update type set delete_flag = '1' where type_id = #{typeId}
+    </update>
+</mapper>
